+++ src/main/java/gov/ca/cwds/idm/persistence/PermissionRepository.java
@@ -3,8 +3,11 @@ package gov.ca.cwds.idm.persistence;
 import gov.ca.cwds.data.auth.ReadOnlyRepository;
 import gov.ca.cwds.idm.persistence.model.Permission;
 import java.util.List;
+
+import org.springframework.context.annotation.Profile;
 import org.springframework.data.jpa.repository.Query;
 
+@Profile("idm")
 public interface PermissionRepository extends ReadOnlyRepository<Permission, String> {
 
   @Query("SELECT p.name FROM Permission p order by p.name asc")
