+++ src/main/java/gov/ca/cwds/idm/service/DictionaryProvider.java
@@ -3,10 +3,12 @@ package gov.ca.cwds.idm.service;
 import gov.ca.cwds.idm.persistence.PermissionRepository;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
+@Profile("idm")
 public class DictionaryProvider {
 
   private PermissionRepository permissionRepository;
