+++ src/test/java/gov/ca/cwds/service/UserAuthorizationServiceTest.java
@@ -34,16 +34,13 @@ public class UserAuthorizationServiceTest {
     AssignmentUnitDao assignmentUnitDao = Mockito.mock(AssignmentUnitDao.class);
     StaffPersonDao staffPersonDao = Mockito.mock(StaffPersonDao.class);
 
-    CwsUserInfoService cwsUserInfoService = new CwsUserInfoService();
-    cwsUserInfoService.setAssignmentUnitDao(assignmentUnitDao);
-    cwsUserInfoService.setCwsOfficeDao(cwsOfficeDao);
-    cwsUserInfoService.setStaffAuthorityPrivilegeDao(staffAuthorityPrivilegeDao);
-    cwsUserInfoService.setUserIdDao(userIdDao);
-    cwsUserInfoService.setStaffUnitAuthorityDao(staffUnitAuthorityDao);
-    cwsUserInfoService.setStaffPersonDao(staffPersonDao);
-
     UserAuthorizationService userAuthorizationService = new UserAuthorizationService();
-    userAuthorizationService.setCwsUserInfoService(cwsUserInfoService);
+    userAuthorizationService.setAssignmentUnitDao(assignmentUnitDao);
+    userAuthorizationService.setCwsOfficeDao(cwsOfficeDao);
+    userAuthorizationService.setStaffAuthorityPrivilegeDao(staffAuthorityPrivilegeDao);
+    userAuthorizationService.setUserIdDao(userIdDao);
+    userAuthorizationService.setStaffUnitAuthorityDao(staffUnitAuthorityDao);
+    userAuthorizationService.setStaffPersonDao(staffPersonDao);
 
     String racfid = "racfid";
     UserId userId = new UserId(null, null, null, STAFF_ID, ID, LOGON_ID, null);
